// src/lib/visual-crossing-schema.d.ts
export interface paths {
// ...
    "/VisualCrossingWebServices/rest/services/timeline/{location}/{startdate}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Historical and Forecast Weather API
         * @description Seamless access to daily and hourly historical and forecast weather data plus weather alerts, events and current conditions.
         */
        get: {
            parameters: {
                query: {
                    /**
                     * @description data format of the output either json or CSV
                     * @example json
                     */
                    contentType?: string;
                    /** @example us */
                    unitGroup?: string;
                    /**
                     * @description data to include in the output (required for CSV format - days,hours,alerts,current,events )
                     * @example days
                     */
                    include?: string;
                    /**
                     * @description Language to use for weather descriptions
                     * @example us
                     */
                    lang?: string;
                    /** @example INSERT_YOUR_KEY */
                    key: string;
                };
                header?: never;
                path: {
                    /** @example London,UK */
                    location: string;
                    /** @example 2022-02-01 */
                    startdate: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Auto generated using Swagger Inspector */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/VisualCrossingWebServices/rest/services/weatherdata/history": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Retrieves hourly or daily historical weather records.
         * @description The weather history data is suitable for retrieving hourly or daily historical weather records.
         */
        get: {
            parameters: {
                query?: {
                    /** @example -1 */
                    maxDistance?: string;
                    /** @example false */
                    shortColumnNames?: boolean;
                    /** @example 2020-02-04T00%3A00%3A00 */
                    endDateTime?: string;
                    /** @example 24 */
                    aggregateHours?: string;
                    /** @example false */
                    collectStationContributions?: boolean;
                    /** @example 2020-01-28T00%3A00%3A00 */
                    startDateTime?: string;
                    /** @example -1 */
                    maxStations?: string;
                    /** @example false */
                    allowAsynch?: boolean;
                    /** @example Sterling%2C%20VA%2C%20US */
                    locations?: string;
                    /** @example false */
                    includeNormals?: boolean;
                    /** @example json */
                    contentType?: string;
                    /** @example us */
                    unitGroup?: string;
                    /** @example INSERT_YOUR_KEY */
                    key?: string;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Auto generated using Swagger Inspector */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}